requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
alembic==1.13.2
pydantic==2.9.1
email-validator==2.1.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
slowapi==0.1.9
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
structlog==24.4.0
python-dotenv==1.0.1